requirements.txt

# Backend
fastapi==0.111.0
uvicorn==0.30.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
python-multipart==0.0.9

# LLM & Agents
langchain==0.2.1
langchain-openai==0.1.7
langchain-community==0.2.1
langchain-google-genai>=2.0.0
llama-index==0.10.43
openai==1.30.1

# Embeddings & Vector DB
sentence-transformers==3.0.0
chromadb==0.5.0
faiss-cpu==1.8.0

# Ingestion
pypdf2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.2
youtube-transcript-api==0.6.2
openai-whisper==20231117

# NLP
nltk==3.8.1
spacy==3.7.4
tiktoken==0.7.0

# Data
numpy==1.26.4
pandas==2.2.2

# Graph
networkx==3.3

# Dev
pytest==8.2.1
httpx==0.27.0
streamlit==1.35.0
streamlit-agraph==0.0.45
